Add a command palette (#80)
This PR adds a basic command palette.

## Changes

- `KeyModifiers` can now retrieve its constituent keys in a human-readable format, ordered by how keybindings are typically written down.
- `Keybind` stores its constituent keys, and overrides `ToString`.
- `VIRTUAL_KEY` has an extendion method to return it as a human-readable string (i.e., without the `VK_` prefix).
- Moved `FocusIndicatorWindow`'s logic for creating a borderless window to an extension method of `Microsoft.UI.Xaml.Window`, in `WindowExtensions`.
- `IWindow` can now focus and force windows to the top of the Z-index, with `FocusForceForeground`.

## Additions

A basic command palette, which has configurable matching (ranking and filters). It takes advantage of the command system added in #79.

The command palette can be activated on a given monitor. Currently, it activates on the currently focused monitor.

## General notes

Performance is lacking in the following scenario:

1. Query the command palette.
2. Delete some characters from the query.

There is a noticeable lag between deleting characters from the query, and the additional rows being shown.