Move PInvoke to be internal to Whim's core (#255)
Add a `public interface INativeManager` to `IConfigContext`, and an `internal interface ICoreNativeManager` to `IConfigContext` for accessing native functionality. These interfaces will wrap `PInvoke` calls, and replace the existing `Win32Helper` static class.

This allows CsWin32 to be `internal` by default and only be exposed to Whim's core. It will also facilitate testing as we are no longer relying on static methods.

`Windows.Win32` structs and enums can be selectively copied to `src/Whim/Native` so they can be exposed outside of the `Whim` assembly.

Also remove `Screen`'s unused `_bitDepth` field, and the `CreateDCW`, `DeleteDC`, and `GetDeviceCaps` methods used to acquire it.