Refactor command palette to support different view variants (#268)
`Whim.CommandPalette` assumed that the types of views showed below the textbox in the command palette would be exclusively free text or a menu. To support #267, this has been refactored so that we can implement variants via the interfaces `IVariantControl` and `IVariantViewModel` to provide additional types of views (like a single/multi-select).

Each view has an associated subclass of `BaseVariantConfig`, which is used to configure the view. This is used to configure the view model, and to provide the view with the necessary information to render the view.

`IPaletteRow` has been replaced with `IVariantItem`, which is a more general interface that can be used to represent any item in the palette. This is used in the menu, and will be used in the single/multi-select.

This PR also marginally improved coverage for `Whim`.