Fix scaling for non-100% monitors (#315)
Previously, `WindowDeferPosHandle` had a static method `SetWindowPosFixScaling` which would call `SetWindowPos`[^1] twice. This was a workaround to fix a single `SetWindowPos` call setting window positions but not their dimensions. `SetWindowPosFixScaling` was called in some of Whim's own windows. Unfortunately, this wasn't applied for `BarWindow`, nor for `Workspace.DoLayout`. As a result, windows would have incorrect dimensions until their second `SetWindowPos` call.

This PR fixes by calling `SetWindowPos` twice when we have a monitor with non-100% scaling, by moving `WindowDeferPosHandle`'s logic into a private class which `WindowDeferPosHandle` calls.

During the course of the investigation, it was discovered that `Monitor.ScaleFactor` was incorrect. This was fixed by manually calcuating the scale by getting the effective DPI using `GetDpiForMonitor` and dividing by 96. This was because `GetScaleFactorForMonitor` would return incorrect values for non-100% scaled monitors.

`BarWindow` was cleaned up a bit by removing logic referring to the scale factor, and reduced some duplication.

[^1]: `SetWindowPos` refers to the entire lifecycle of `BeginDeferWindowPos` to `EndDeferWindowPos`, not the `SetWindowPos` function itself.