Show tree bar widget on all relevant bars (#324)
The `TreeLayoutEngineWidgetViewModel` now updates correctly when receiving events from `WorkspaceManager`. Previously it would only populate on first load if the bar was in the focused monitor. However, we want each bar to show the state for their monitor, not for the active workspace.

To fix this, we are now cognizant of the monitor when receiving updates.