`TreeLayoutWidget` now listens to changes from the `TreeLayoutPlugin` (#330)
`TreeLayoutPlugin` now notifies subscribers of changes in the current `AddNodeDirection` via the event `AddWindowDirectionChanged`. This allows `TreeLayoutWidget` now listens to changes from the `TreeLayoutPlugin`, and resolves the issue where the `TreeLayoutWidget` would not update when the shown direction was not changed via the command palette.

`Whim.TreeLayout` now has significantly reduced API surface, and the recommended way of interacting with `TreeLayoutEngine` is via the plugin. Asaa result of this the `BarConfig` was updated to allow the component lists to be changed up until Whim being initialized.

`TreeLayoutEngineWidget` has been simplified to rely on the `TreeLayoutPlugin` for all of its data - it no longer maintains any internal state indicating the current direction.

The default `whim.config.csx` has been updated to reflect the changes to the `Whim.TreeLayout.Bar`.