Refactor command and keybind managers (#340)
Previously, `ICommandManager` provided the interface for managing keybinds. This was highly coupled. Instead, `IKeybindManager` has been reintroduced to manage keybindings. This also lets us have a single keybind have multiple commands bound to it. This is useful because commands can be conditional/situational.

`IPlugin` has been updated to use a new `PluginManager` class, which manages the commands and keybinds provided by a plugin. While I did consider in #299 allowing commands to not be added, I decided against this for simplicity's sake, and because Visual Studio Code behaves the same way. 

The `PluginManager` has a nicer API for defining commands and keybinds, and it also allows for commands to be defined in a more declarative way. For example, the following code:

```csharp
Add(
    identifier: "toggle",
    title: "Toggle command palette",
    _commandPalettePlugin.Toggle,
    keybind: new Keybind(IKeybind.WinShift, VIRTUAL_KEY.VK_K)
)
.Add(
    identifier: "rename_workspace",
    title: "Rename workspace",
    callback: () =>
        _commandPalettePlugin.Activate(
            new FreeTextVariantConfig()
            {
                Callback = (text) => _context.WorkspaceManager.ActiveWorkspace.Name = text,
                Hint = "Enter new workspace name",
                InitialText = _context.WorkspaceManager.ActiveWorkspace.Name,
                Prompt = "Rename workspace"
            }
        )
);
```

Each command id will automatically be prefixed by the plugin's name. For example, the two commands in the above example would have ids `whim.command_palette.toggle` and `whim.command_palette.rename_workspace`. This helps identify

`IPlugin` commands and keybinds are now loaded in `PluginManager.AddPlugin`. This guarantees that user-defined keybinds can override plugin keybinds as long as they are defined in the `IKeybindManager` _after_ the plugin is loaded by the `PluginManager`.

Coverage was increased, and some minor refactoring to support this was performed.