Save plugin state to JSON (#347)
Plugins can now save their state to JSON in `~/.whim/state`. Plugins should be careful what they save, and that interacts with any config the user has set.

The `ActivationConfig` of the `CommandPaletteConfig` is now of type `MenuVariantConfig`, and cannot be overridden by the user. However, the matcher in the `ActivationConfig` can be changed.

This PR doesn't add save any state of Whim's core functionality. It seemed like a gnarly can of worms to open, as it's tricky to track windows when Whim isn't running. What if there are two windows with the same name, but in different workspaces? How do we know which to put where? 

In the end, the config should suffice for most cases. Saving state was mainly added so the command palette can store its weighting for commands.