Add improved `FocusIndicator` reactivity (#395)
The focus indicator is now more reactive. Previously, it would not update when a window's size had changed or the active layout engine had changed. Now, it will update in these cases.

This was achieved by adding events for when `IWorkspace.DoLayout` has started and completed, and subscribing to these events in the `FocusIndicatorPlugin`. Additionally, `WindowDeferPosHandle` now accepts custom `SET_WINDOW_POS_FLAGS`. This prevents flashing in most cases when the focus indicator is updated.