Replace public `Trigger*` methods with `WorkspaceManagerTriggers` (#397)
This replaces the internal `Trigger` methods with a `Trigger` method with a `WorkspaceManagerTriggers` class, which contains references to `WorkspaceManager` methods which triggers events. The `WorkspaceManagerTriggers` are passed to the `Workspace` constructor.

As a result of this change, it is no longer possible to create a `Workspace` methods outside of the `WorkspaceManager`. Instead, it's possible to set the property `CreateLayoutEngines`, a function which creates an array of layout engines to use when creating workspaces. `CreateLayoutEngines` replaces `WorkspaceFactory`.