Switch to moving fractions based on pixel deltas (#416)
Previously, Whim would move window edges based on a fraction of the overall monitor size. This PR switches to moving window edges based on a pixel size. This was done in preparation for #411. Additionally, from a user perspective, it makes sense to work on a pixel basis - it's consistent across monitors.