Resize windows by dragging their edges (#418)
This PR adds support for resizing windows by dragging their edges or corners. Implemention of this is dependent on the underlying `ILayoutEngine`.

To achieve this, `MoveWindowEdgeInDirection` was renamed to `MoveWindowEdgesInDirection`, and accepts an additional parameter `IPoint<int>` to handle corner resizing.

The enum `Direction` now has the `Flags` annotation, and bitwise operations are used to create values for diagonal directions.

Coverage was significantly improved for `WindowManager`. To facilitate this increase in coverage, `MonitorManager` now implements a new interface `IInternalMonitorManager`. Similarly, `WorkspaceManager` implements a new interface `IInternalWorkspaceManager`. This makes it easy to mock out methods which are Whim `internal`-only.