Remove old windows (#502)
At the start of `DoLayout`, before starting to call `DeferWindowPos`, Whim now checks that each window in the workspace identifies an existing window. If the window does not, Whim removes the window from itself. Additionally, the `DeferWindowPos` call is cancelled, relying instead on the subsequent `DoLayout` call after the window has been removed.