Move `SetWindowPos` into a different thread (#505)
Previously, double clicking a focused window from the taskbar would cause the window to minimize and restore, triggering a layout loop. To fix this, Whim now sets the layouts of windows in a separate thread. This means that Whim can respond to window events more quickly, as it doesn't have to wait for the layout to finish before responding to a new event.

There was a race condition in `WorkspaceWidgetViewModel`. This has been fixed by simplifying the logic for setting `WorkspaceModel.ActiveOnMonitor`.