Fix workspace activation race condition (#510)
Added a `ReaderWriterLockSlim` to prevent `WindowManager` from modifying `MonitorManager` and `WorkspaceManager` while a `Workspace.DoLayout` is occurring in a different thread. `Reader` locks are used for `Workspace.DoLayout`, while `Writer` locks are used for `WindowManager` methods which modify `MonitorManager` and `WorkspaceManager`. This was done to fix a race condition between `Workspace.DoLayout` being called on the wrong monitor after `MonitorManager.WindowFocused` was called.

`WindowDeferPosHandle` was simplified to reduce the allocations and instantiations, by removing the `InternalWindowDeferPosHandle` and accepting an enumerable of windows to layout.

`Workspace` now uses a queue to cancel all prior layouts which are currently in progress.

`IWorkspace` now (regrettably) returns `Task` for `DoLayout` and methods which call `DoLayout` (e.g. `AddWindow`). This is to allow the caller to `await` the layout operation. However, it would be nice to remove all `Task.Run` calls in the future, for simplicity's sake.