Add `INotificationManager` to facilitate toast notifications (#721)
Added `INotificationManager` to facilitate toast notifications. The core and plugins can display notifications and register listeners to handle user interaction with the notifications. This is a bit of a tricky API, as _all_ interactions need to set the `WHIM_NOTIFICATION_ID` argument in order for Whim to route the interaction to the correct listener.

This will be used by the `Whim.Updater` in #141.