Let `ILayoutEngine`s perform custom actions  (#736)
This PR adds the ability for layout engines to perform non-standard actions, and still have their resulting layouts be saved to `Workspace` instances. This was needed for #634, where the `SliceLayoutPlugin` tries to perform operations on the `SliceLayoutEngine`, but was unable to save the resulting `ILayoutEngine` instance to the `Workspace`.

The `PerformCustomAction` method can be used by `ILayoutEngine` implementations to perform non-standard actions.