Add set/unset maximize actions to focus engine (#762)
Currently, the FocusLayoutEngine API does not expose the state of `_maximized`. It is hence not possible to conditionally use `toggle_maximize` to set a certain state along the following lines:
```python
if not _maximized:
    toggle_maximized
```

This PR adds two new custom actions that allow setting the `_maximized` state explicitly. For now, I haven't added them to the core commands to avoid pollution.

---------

Co-authored-by: Isaac Daly <isaac.daly@outlook.com>