Try avoid moving windows when removing/adding monitors (#824)
When Windows removes/adds monitors, it will helpfully move windows to/from monitors. However, Whim is a window manager and we'd like to control this behavior.

To work around Windows' helpfulness, the `WindowManager` and `ButlerEventHandlers` ignore all Windows events for tracked windows. This has the downside of potentially missing a window closing, but I've judged that it's more important to keep windows in the correct workspaces.

Other changes include:

- Removing an ineffectual lock
- Updating some calls to obsolete methods
- Another attempt at avoiding `AccessViolationException` in the `Bar`