Fix bar alignment (#837)
Small fix and clean up of the bar plugin:

1. Remove unused `Grid.Column` attached properties 
2. Fix alignment of center panel

## Current issue

The `CenterPanel` is centered within the space _unoccupied_ by the other two panels. If the `LeftPanel` and `RightPanel` have different widths, this results in the `CenterPanel` being off-centered. 

Moreover, if the width of one of the other panels changes, the `CenterPanel` will move around. This occurs for example if the `RightPanel` contains the `FocusedWindowWidget`.

## Diagnosis

The issue is caused because:

> The panel-center alignment properties ([AlignVerticalCenterWith](https://learn.microsoft.com/en-us/uwp/api/windows.ui.xaml.controls.relativepanel.alignverticalcenterwith), [AlignHorizontalCenterWithPanel](https://learn.microsoft.com/en-us/uwp/api/windows.ui.xaml.controls.relativepanel.alignhorizontalcenterwithpanel), ...) are typically used independently of other constraints and are applied if there is no conflict.

In other words, `AlignHorizontalCenterWithPanel` is ignored due the presence of the sibling positional relations. (The reason it still looks roughly centered, is that `HorizontalAlignment="Center"` causes it to center _within its available space_).

## Proposed fix

Eliminating the sibling positional relations from the `CenterPanel` does the trick. The three panels are still ordered correctly based on their absolute positions. (The left panel defaults to `AlignLeftWithPanel` -- we could make that explicit)