Update komorebi application rules license (#951)
A tiny update to `THIRD-PARTY-NOTICES`, reflecting the newly added [MIT license](https://github.com/LGUG2Z/komorebi-application-specific-configuration/blob/master/LICENSE) to the komorebi-application-specific-configuration collection.

In the past, we had included the license of the main komorebi repo instead (which used to be MIT as well). So this ends up just being a change in date.