Always expose the workspace id when adding a workspace (#977)
Previously, `context.WorkspaceManager.Add()` would return `null` in the csx. This was because it would only return `IWorkspace` once Whim had completed initialization.

Now, `context.WorkspaceManager.Add()` will return the `Guid` of the workspace that was added. This allows you to utilize the workspace id when setting up commands or other functionality that requires the workspace id.