C# exceptions are too slow (#1131)
This pull request replaces DotNext's `Result` type with a new `Result` type from within Whim itself. The reason for this was that `DotNext.Result` relied on `Exception`s for error handling, or an `Error` enum. `Exception`s are slow (can take up to a second to create) and are not suitable for performance-sensitive code. The new `Result` type maintains much of the same API, but accepts a `WhimError` type for errors, which includes a `string` message and an optional `Exception` for library code which provide any errors.

`DotNext` has been removed as a dependency.